requirements.txt
streamlit
Pillow
pandas
streamlit-image-coordinates
pymupdf
